Included complete brokerage_cust dataset, Client page rendering all data Changed client api route to /api/clients
